refactor(sampling): move sampling rules and configuration to their own files
