+++ datadog-opentelemetry/src/core/configuration/mod.rs
@@ -22,10 +22,10 @@
 #[allow(clippy::module_inception)]
 mod configuration;
 pub(crate) mod remote_config;
+mod sampling_rule_config;
 mod sources;
 mod supported_configurations;
 
-pub use configuration::{
-    Config, ConfigBuilder, OtlpProtocol, SamplingRuleConfig, TracePropagationStyle,
-};
+pub use configuration::{Config, ConfigBuilder, OtlpProtocol, TracePropagationStyle};
 pub(crate) use configuration::{ConfigurationProvider, RemoteConfigUpdate};
+pub use sampling_rule_config::SamplingRuleConfig;
