+++ datadog-opentelemetry/src/sampling/mod.rs
@@ -11,12 +11,14 @@ pub(crate) mod otel_mappings;
 pub(crate) mod rate_limiter;
 pub(crate) mod rate_sampler;
 pub(crate) mod rules_sampler;
+pub(crate) mod sampling_rule;
 mod types;
 pub(crate) mod utils;
 
 // Re-export key public types
-pub use datadog_sampler::{DatadogSampler, SamplingRule, SamplingRulesCallback};
+pub use datadog_sampler::{DatadogSampler, SamplingRulesCallback};
 pub use otel_mappings::{OtelAttributeFactory, OtelSamplingData};
+pub use sampling_rule::SamplingRule;
 pub use types::{
     AttributeFactory, AttributeLike, SamplingData, SpanProperties, TraceIdLike, ValueLike,
 };
